Refactor active chat conversation loading
- extract active conversation fetch, hydration, highlights, socket join/read marking, and pagination into useActiveChatConversation
- keep empty direct chats transient and filter persistent empty DM rows
- scope conversation query cache per authenticated user
- update chat refactor docs and README
